migrate main prerender suite
