+++ test/e2e/prerender/pages/api/bad.js
@@ -1,0 +1,3 @@
+export default function handler(req, res) {
+  throw new Error('lol')
+}
