+++ test/e2e/prerender/pages/bad-gssp.js
@@ -1,0 +1,7 @@
+export function getServerSideProps() {
+  throw new Error('lol')
+}
+
+export default function BadGssp() {
+  return <div />
+}
