+++ test/e2e/prerender/pages/bad-ssr.js
@@ -1,0 +1,7 @@
+export function getServerSideProps() {
+  return { props: {} }
+}
+
+export default function BadSsr() {
+  throw new Error('oops')
+}
