+++ test/e2e/prerender/pages/lang/[lang]/about.js
@@ -1,0 +1,12 @@
+export default ({ lang }) => <p id="about">About: {lang}</p>
+
+export const getStaticProps = ({ params: { lang } }) => ({
+  props: {
+    lang,
+  },
+})
+
+export const getStaticPaths = () => ({
+  paths: ['en', 'es', 'fr', 'de'].map((p) => `/lang/${p}/about`),
+  fallback: false,
+})
