+++ test/e2e/prerender/pages/normal.js
@@ -1,0 +1,1 @@
+export default () => <p id="normal-text">a normal page</p>
