+++ test/e2e/prerender/pages/user/[user]/profile.js
@@ -1,0 +1,28 @@
+import React from 'react'
+import Link from 'next/link'
+
+export async function getStaticPaths() {
+  return { paths: [], fallback: true }
+}
+
+export async function getStaticProps({ params }) {
+  return {
+    props: {
+      user: params.user,
+      time: (await import('perf_hooks')).performance.now(),
+    },
+    revalidate: 10,
+  }
+}
+
+export default ({ user, time }) => {
+  return (
+    <>
+      <p>User: {user}</p>
+      <span>time: {time}</span>
+      <Link href="/">
+        <a id="home">to home</a>
+      </Link>
+    </>
+  )
+}
