+++ test/e2e/prerender/world.txt
@@ -1,0 +1,1 @@
+world
